Add files via upload
This is the code I used whilst writing report. The Graph & Gui file is code that is written by a colleague, the rest is my own.

This contains two examples of how a map could be displayed. The path-finding algorithm is built upon a Dijkstra’s Algorithm library in racket, modified for use here. You can see more about how all this works within the report.