Add files via upload
This is the code used for the project. The Graph & Gui code is written by a colleague, the rest being my own. 

The path-finding algorithm is built off a racket Dijkstra’s Algorithm library that is modified for use. It's paired with an example GUI. There are two other examples displaying how a map could be displayed. 

You can read more about this within the report.